refactor(docs): separate texts components in a section
