+++ src/pages/ui-components/Texts/Subtitle/index.tsx
@@ -4,33 +4,33 @@ import { IStyledProp } from '@eduzz/houston-styles/styled';
 import useHoustonTheme from '@eduzz/houston-styles/useHoustonTheme';
 import type { HoustonTokens } from '@eduzz/houston-tokens';
 
-import useMediaQuery from '../hooks/useMediaQuery';
+import useMediaQuery from '../../hooks/useMediaQuery';
 import Typography from '../Typography';
 
-type ParagraphTags = 'p' | 'span' | 'strong';
+type SubtitleTags = 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'p';
 
-export type ParagraphSizes = 'large' | 'small';
+export type SubtitleSizes = 'large' | 'small';
 
 export interface IHeadingProps extends IStyledProp {
   id?: string;
   /**
    * Defaults to 'large'
    */
-  size?: ParagraphSizes;
+  size?: SubtitleSizes;
   weight?: keyof HoustonTokens['font']['weight'];
   lineHeight?: keyof HoustonTokens['line']['height'];
   marginBottom?: boolean;
   children?: React.ReactNode;
   onClick?: (e: React.MouseEvent) => void;
   /**
-   * Defaults to 'p'
+   * Defaults to 'h2'
    */
-  as?: ParagraphTags;
+  as?: SubtitleTags;
   ['aria-label']?: string;
 }
 
 type SizesMap = {
-  [key in ParagraphSizes]: keyof HoustonTokens['font']['size'];
+  [key in SubtitleSizes]: keyof HoustonTokens['font']['size'];
 };
 
 const defaultSizesMap: SizesMap = {
@@ -39,12 +39,12 @@ const defaultSizesMap: SizesMap = {
 };
 
 const mobileSizesMap: SizesMap = {
-  large: 'sm',
-  small: 'xs'
+  large: 'md',
+  small: 'sm'
 };
 
-const Paragraph = React.forwardRef<HTMLParagraphElement | HTMLSpanElement | HTMLElement, IHeadingProps>(
-  ({ as = 'p', children, size: sizeProp = 'large', ...props }, ref) => {
+const Subtitle = React.forwardRef<HTMLHeadingElement | HTMLParagraphElement, IHeadingProps>(
+  ({ as = 'h2', children, size: sizeProp = 'large', ...props }, ref) => {
     const { breakpoints } = useHoustonTheme();
     const isMobile = useMediaQuery(breakpoints.down('sm'));
     const sizesMap = isMobile ? mobileSizesMap : defaultSizesMap;
@@ -57,4 +57,4 @@ const Paragraph = React.forwardRef<HTMLParagraphElement | HTMLSpanElement | HTML
   }
 );
 
-export default Paragraph;
+export default Subtitle;
